Add tutorial section
